requirements.txt
fastapi
uvicorn
pydantic
openai
streamlit
requests
google-generativeai
googletrans